msi: change when a prior version is uninstalled
This commit changes the default value of the `@Schedule` attribute of the
`<MajorUpgrade />` element so that if a newer version of Git for Windows
fails to install on upgrade, the uninstallation of the previous version
will be rolled back. This change alone is one of the major motivations
behind issue #682 to begin with.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>