msi: the user's core.autocrlf choice is configured
The MSI installer now respects the users' chosen `core.autocrlf` setting
during installation and configures it in the git system configuration file
for MSYS2.

The setting will be appropriately configured whether the installation is
performed as a full UI install, or a reduced UI or silent installation.
The setting is appropriately configured during repair as well.

When this setting is applied on the command line, the values are accepted
case-insensitively. This is also important because the user _could_
manually change the registry value for this option, changing the casing,
which would then break an upgrade installation. (Property values are case-
sensitive, especially when they are used to back a Radio Button group,
such as this option. So the installer translates the value of LINEENDINGS
to the appropriate casing so that the UI can display the chosen option
correctly.)

This property follows the Remember Property pattern while allowing a
default value to exist for the option. More information about the
Remember Property pattern can be found at [Rob Mensching's
blog](http://robmensching.com/blog/posts/2010/5/2/the-wix-toolsets-remember-property-pattern/).
The main difference is that instead of saving any command line value
received for the property, the command line value is saved only if it
differs from the default value. The Remember Property pattern, as written,
assumes the property does not have a (default) value (which means that it
would not be authored into the MSI Property table). Since this property
has a default value, if a command line value is not provided, the property
will retain it's authored value and be saved as if it had been provided on
the command line. Then, if the user is upgrading, the registry is searched
for the property value. If the installed value differs from the default,
the property value is updated. But later, the saved "command line" value
is "restored" (which wasn't really a command line value, but the property
default value), effectively disregarding the user's prior installation
setting. By only saving the command line value if it differs from the
default, this limitation of the default implementation of the Remember
Property pattern is overcome.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>