msi: add SetProperty condition and fix whitespace
Since the custom action to run `post-install.bat` will not run if the user
is completely removing Git for Windows from their system, there's no need
to set the property that sets the command to be run to execute the batch
file. This commit adds a condition to not execute this custom action on
removal to reduce the amount of time it takes to remove Git for Windows
from the system. (NOTE: Git for Windows will be removed from the system on
upgrades, so this mode  will run more frequently than just the case where
the user is explicitly uninstalling Git for Windows.)

Lastly, collapsed the definition of the custom action to execute the
`post-install.bat` file to one line.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>