shears: work around patches merged upstream
When rebasing a branch that has been applied upstream, the shears will
get confused because they won't find the tip commit of said branch in
the todo script.

Simply skip this for now; If this developer gets his way, the shears
functionality will eventually enter the sequencer, with a lot better
control than any shell script can hope for.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>