Move Makefile, edit-git-bash.c to build-extra
The InnoSetup installer requires `edit-git-bash.c` to be compiled to a
DLL, though currently, this is not compiled as part of the overall release
process. This DLL is used by the installer to modify the `git-bash.exe`
string resources responsible for starting the terminal emulator for hosting
the BASH. It comes into play when the installing user chooses to use the
standard Windows Command Console over the default mintty terminal emulator.

The new MSI installer also needs a way to modify `git-bash.exe` string
resources to achieve the same effect. The easiest way to accomplish this
is to compile `edit-git-bash.c` as an EXE and use a custom action to call
the `edit-git-bash.exe` to modify `git-bash.exe`.

Therefore, this commit moves the make file and `edit-git-bash.c` to the
top-level `build-extra` directory since it's a resource now required by
more than one installer.

Likewise, the .gitignore has been updated to ignore `edit-git-bash.dll`
and `edit-git-bash.exe`.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>