WIP: innosetup: support optional packages
This is a first crack at adding support for including additional packages
in the InnoSetup distribution vehicle. It's rough and contains lots of
commented out code and debug output.

All other release vehicles (msi, nuget, portable) allow you to include
additional, optional packgaes, but InnoSetup release process does not.

This developer is reluctant to make this feature addition lightly because
this is the primary release vehicle for Git for Windows.

So, this developer is putting this commit out there for any early feedback
and to perhaps at least spark a discussion, if nothing else.

This change was instigated via
https://github.com/git-for-windows/git/issues/771.