innosetup: build edit-git-bash.dll in release.sh
The Makefile was modified to ensure that `edit-git-bash.c` is always
compiled as a 32-bit DLL when the `edit-git-bash.dll` target is invoked.
InnoSetup is only 32-bit "aware", so even for the 64-bit distribution, a
32-bit DLL needs to be used.

The InnoSetup `release.sh` script was updated to include building
`edit-git-bash.dll`. The rendering of release notes was moved until after
all argument processing; this also had the advantage of placing the call
to `make` directly after it--these activities are sort of "pre-processing"
before the actual generation of the installer and made logical sense to
this developer.

Lastly, the `install.iss` file was updated to source the DLL from it's new
location in `build-extra`; the old edit-git-bash.dll in the `installer`
directory was deleted.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>