msi: make all post-install/*.post files unmanaged
When the MSI was first added to the code base, `post-install.bat` was made
a MSI unmanaged component by setting it's component Guid="". This was
because `post-install.bat` deletes itself.

In a similar fashion, the `post-install.bat` file ensures all the
`etc/post-install/*.post` files get executed, including
`99-post-install-cleanup.post`, which ensures all the `*.post` files are
removed. Therefore, the `*.post` files should also be made unmanaged,
since they are "installed" by the installer, only to be deleted during
installation.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>