Add debug output to edit-git-bash
While implementing the option to choose which terminal emulator to use
when installing Git for Windows via MSI, it came in really handy to be
able to see the full command line call to `edit-git-bash.exe` and the
arguments as parsed by a call to `CommandLineToArgvW(...)`. So, wrap this
debugging output in a `#ifdef DEBUG` block.

This output will be shown in the MSI logs when logging verbose output and
can help track down any weirdness with setting the string resources in
git-bash.exe.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>