installer: detect and discourage downgrades
It is too easy to hard-code URLs to specific Git for Windows versions'
installers. That means that any 3rd-party software that wants to install
Git for Windows runs the danger of forcing Git for Windows to be
downgraded.

This is obviously not what we want, nor what the users want, so let's
show a message box and ask the user whether they want to downgrade,
defaulting to "No".

Silent installs will silently prevent downgrades.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>