msi: fix dialog flow for line endings
The first implementation of the UI always showed a dialog asking if you
want to customize how line endings are handled by Git for Windows, or just
use the default. In most cases, those who don't know what setting to
choose should leave well enough alone.

However, for those individuals who know the consequences of using a
non-default value, this dialog is just a nuisance and can, in its original
implementation, be confusing as to the value that will actually be used
for installation.

So, for a fresh install, don't show this dilaog if a value for
`LINEENDINGS` was set to a value other than `true` on the command line to
`msiexec`. In this case, the user is passing a non-default option and
presumably knows what they're doing. Don't get in their way.

On an upgrade, don't show this dialog if the previous installation used a
value other than `true` OR the upgrade installation was invoked via the
command line passing a value other than `true` for `LINEENDINGS`.

In all other cases, the dialog will still be shown.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>