msi: fix dialog flow for line endings
The first implementation of the UI always showed a dialog asking if you
want to customize how line endings are handled by Git for Windows, or just
use the default. In most cases, those who don't know what setting to
choose should leave well enough alone.

However, for those individuals who know the consequences of using a
non-default value, this dialog is just a nuisance and can, in its original
implementation, be confusing as to the value that will actually be used
for installation.

This commit changes the flow of the UI dialogs pertaining to choosing the
line endings in Git for Windows. This was accomplished by introducing a
new variable, `SkipGitLineEndingsDlg`. This variable will be set to `Yes`
if any one of the following conditions is true, and will suppress the
display of the `GitLineEndingsDlg` and instead advance directly to the
`GitCustomizeLineEndingsDlg` with the appropriate radio button selected:

- If a valid value is specified for `AUTOCRLF`
- If an upgrade, repair, or reinstall is taking place, and the value
  chosen during the previous installation was not `true`.
    - NOTE: If the registry contains an invalid value for this setting
      (e.g. a user manually set this registry entry to an invalid value),
      the value will default to `true` and `SkipGitLineEndingsDlg` will be
      `No`.

Introducing the `SkipGitLineEndingsDlg` variable also makes it easier to
implement `Back` button functionality to properly determine when to skip
the `GitLineEndingsDlg` as well.

Additionally, the dialog flow was unintuitive when clicking the Back
button, and could cause an unintended configuration to be used. Here's the
sequence of steps that would lead to the problem:

1. A user selects "Yes, customize the line endings.
2. The user is presented with options for selecting how Git for Windows
   should handle line endings.
3. The user selects a non-default value (e.g. `input` or `false`).
4. The user advances through the installer.
5. At some point, the user clicks the `Back` button all the way back to
   the `GitLineEndingsDlg` where they can choose whether or not they want
   to customize their line endings.
6. The user changes their mind and now selects "No, don't customize line
   endings."
7. The user clicks Next.

In this workflow, after clicking `Next`, the user is taken to the Choose
Terminal dialog because they indicated they don't want to customize line
endings. And, because the user selected "No, don't customize line
endings", the user would (rightly) assume that the default value for
line endings would be used (`true`). However, that was not the case; the
value for `LINEENDINGS` would not be changed on the `GitLineEndingsDlg`,
but instead it would retain the previously customized value. And so it is
possible that the user would end up with a configuration setting they did
not want nor intend.

This commit fixes that flow so that if the user ever selects "No, don't
customize line endings" on the `GitLineEndingsDlg` and subsequently presses
either the `Back` or `Next` buttons, the `LINEENDINGS` property value is
updated to `true`, the default value for Git for Windows. NOTE that the
property value is changed only if the user clicks `Back` or `Next`. This
way, if the user changes their mind and clicks "No, don't customize", but
then changes their mind yet again before leaving the dialog, the previous
setting _would_ be retained and they could then change that setting on the
next dialog.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>