msi: fix TERMINAL property
While initially implementing this MSI property, this developer made an
assumption that the default value was initially used for an installation
in order to test that upon ugrade the property would be set to the
original installation value. In addition, this developer tested setting
the property on the command line during upgrade scenarios--again, under
the assumption that the default value was used for the initial
installation.

Unfortunately, this developer never took the implementation to the next
step, where the initial installation used a non-default value in order to
test that the installer still behaved appropriately.

The original implementation used the "Remember Property" pattern from
@robmen's blog. However, after much trial and error and careful analysis
of the pattern, this developer has come to the conclusion that, contrary
to the blog article, this pattern WILL NOT work for properties which have
default values. This is because when a property is set on the command
line, it replaces the original property value as defined in the
installer's Property table. However, there's no way to know that this
happened, and so for all you know (inside the installer), the property
_always_ had that value. In other words, there's no way to know where the
property value comes from. And this can cause problems with UI dialog flow
and being able to correctly support upgrade and repair scenarios (either
with or without a UI).

This commit fixes the `TERMINAL` property so that upgrade scenarios work
correctly, especially when the installed value for the property is
initially a non-default value _and_ the property value is overridden on
the command line for one of these installation scenarios.

NOTE: This does not apply for repair/reinstall scenarios because during
repair/reinstallation, command line values for public properties are
ignored by Windows Installer; furthermore, during repair/reinstallation,
the original installation values should be used due to the way Windows
Installer performs component tracking. Of course, I can't control what a
user does with the installed registry values--but that's not my problem.

NOTE: If a user were to change the installed registry value from
`CmdPrompt` to `MinTTY`, the user would be sadly disappointed anyway, as
the component used to update `git-bash.exe` string resources so that
`bash` is run from within the standard Windows Command Prompt window
instead of MinTTY only knows how to perform string resource conversion in
a single direction. In any event, this behavior only reinforces that if
the user wishes to change this installation option, he/she must uninstall
and then install Git for Windows again, choosing the appropriate
installation option.

The implementation for getting the user's `TERMINAL` choice works as
follows:

1. A variable is created specifically for use on the command line
   (`CONSOLE`)

2. A private variable, `GfwTerminal` contains the default value for
   this configuration option (`MinTTY`).
    - This property is a "bookkeeping" property that tracks the final
      value of the `TERMINAL` property (discussed next)--unless a Full UI
      installation is occurring (which is most likely the case), in which
      case this value could be overridden in the UI.

3. The `TERMINAL` property is the property associated with the radio
   buttons selection choices in the UI--therefore, it must be initialized
   to one of the valid radio button values. Naturally, it's set to the
   default value for the configuration option, `MinTTY`.
    - This property is public (because it's in all CAPS), and so
      technically it could be set on the command line. However, the
      implementation ensures that this property is always set to the value
      of `GfwTerminal` prior to commencing the UI or the installation (if
      not performing a Full UI installation).

4. A `TERIMNALSEARCH` variable is created and is populated by searching
   the registry for a string-valued entry named `Terminal`.

5. The `GfwTerminal` property value is updated to the value stored
   in `TERMINALSEARCH` if an upgrade, repair, or reinstall is taking place
   _and_ if it contains a valid value found in the registry.

6. The `GfwTerminal` property value is updated again to the value stored
   in `CONSOLE`, which would be set by a user when invoking the installer
   on the command line. The `CONSOLE` property must contain a valid value
   in order for this assignment to occur. NOTE: During reinstall/repair
   scenarios, since Windows Installer ignores command line property
   values, `CONSOLE` will have the value `Unset`, which is not a valid
   value. And therefore, there's no chance that `GfwTerminal` will be
   updated, preserving the value retrieved from `TERMINALSEARCH`, ensuring
   that repairs/reinstall are not affected.

At this point, determining the value for the configuration option is
finished--`GfwTerminal` contains the value that should be displayed
in the UI (or which will be used during installation if not performing a
full UI install). The steps above ensure that the configuration option
obeys the order of precedence of setting this configuration value (in
order of highest precedence to lowest precedence):

1. The value selected through the installer UI (when performing a full UI
   install for the first time or during an upgrade).
2. The value provided to `CONSOLE` when invoking the installation from
   the command line (only during first-time or upgrade installations).
3. The value chosen during a previous installation--i.e. an upgrade,
   repair, or reinstallation is taking place.
4. The default value for the configuration option.

The rest of the implementation makes sure that the registry value
(retrieved during upgrade) and the value set on the command line for
`CONSOLE` have the proper casing--the radio button values in the UI are
case-sensitive. Finally, `TERMINAL` is unconditionally set to the value of
`GfwTerminal`. This ensures that even if a user attempts to circumvent the
steps above, `TERMINAL` always contains a value based on the order of
precedence listed above.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>