msi: fix ENABLEFSCACHE property
While initially implementing this MSI property, this developer made an
assumption that the default value was initially used for an installation
in order to test that upon ugrade the property would be set to the
original installation value. In addition, this developer tested setting
the property on the command line during upgrade scenarios--again, under
the assumption that the default value was used for the initial
installation.

Unfortunately, this developer never took the implementation to the next
step, where the initial installation used a non-default value in order to
test that the installer still behaved appropriately.

The original implementation used the "Remember Property" pattern from
@robmen's blog. However, after much trial and error and careful analysis
of the pattern, this developer has come to the conclusion that, contrary
to the blog article, this pattern WILL NOT work for properties which have
default values. This is because when a property is set on the command
line, it replaces the original property value as defined in the
installer's Property table. However, there's no way to know that this
happened, and so for all you know (inside the installer), the property
_always_ had that value. In other words, there's no way to know where the
property value comes from. And this can cause problems with UI dialog flow
and being able to correctly support upgrade and repair scenarios (either
with or without a UI).

This commit fixes the `ENABLEFSCACHE` property so that upgrade scenarios
work correctly, especially when the installed value for the property is
initially a non-default value _and_ the property value is overridden on the
command line.

NOTE: This does not apply for repair/reinstall scenarios because during
repair/reinstallation, command line values for public properties are
ignored by Windows Installer; furthermore, during repair/reinstallation,
the original installation values should be used due to the way Windows
Installer performs component tracking. Of course, I can't control what a
user does with the installed registry values--but that's not my problem.

The implementation for getting the user's `ENABLEFSCHOICE` choice works as
follows:

1. A variable is created specifically for use on the command line
   (`USEFSCACHE`)

2. A private variable, `GfwEnableFsCache` contains the default value for
   this configuration option (`true`).
    - This property is a "bookkeeping" property that tracks the final
      value of the `ENABLEFSCACHE` property (discussed next)--unless a
      Full UI installation is occurring (which is most likely the case),
      in which case this value could be overridden in the UI.

3. The `ENABLEFSCACHE` property is the property associated with the check
   box on the `GitAdditionalOptionsDlg`.
    - This property is public (because it's in all CAPS), and so
      technically it could be set on the command line. However, the
      implementation ensures that this property is always set to the value
      of `GfwEnableFsCache` (or cleared) prior to commencing the UI or the
      installation (if not performing a Full UI installation).

4. A `FSCACHESEARCH` variable is created and is populated by searching the
   registry for a string-valued entry named `core.fscache`.

5. The `GfwEnableFsCache` property value is updated to the value stored
   in `FSCACHESEARCH` if an upgrade, repair, or reinstall is taking
   place _and_ if it contains a valid value found in the registry.

6. The `GfwEnableFsCache` property value is updated again to the value
   stored in `USEFSCACHE`, which would be set by a user when invoking the
   installer on the command line. The `USEFSCACHE` property must contain a
   valid value in order for this assignment to occur. NOTE: During
   reintsall/repair scenarios, since Windows Installer ignores command
   line property values, `USEFSCACHE` will have the value `Unset`, which
   is not a valid value. And therefore, there's no chance that
   `GfwEnableFsCache` will be updated, preserving the value retrieved from
   `FSCACHESEARCH`, ensuring that repairs/reinstalls are not affected.

At this point, determining the value for the configuration option is
finished--`GfwEnableFsCache` contains the value that should be displayed
in the UI (or which will be used during installation if not performing a
full UI install). The steps above ensure that the configuration option
obeys the order of precedence of setting this configuration value (in
order of highest precedence to lowest precedence):

1. The value selected through the installer UI (if performing a full UI
   install for the first time or during an upgrade).
2. The value provided to `USEFSCACHE` when invoking the installation from
   the command line (only during first-time or upgrade installations).
3. The value chosen during a previous installation--i.e. an upgrade,
   repair, or reinstallation is taking place.
4, The default value for the configuration option.

The rest of the implementation makes sure that the registry value
(retrieved during upgrade, repair, and reinstalls) and the value set on
the command line for `USEFSCACHE` are normalized to one of either `true`
or `false` (since Git configuration booleans can accept a number of
aliases that achieve the same effect). `ENABLEFSCACHE` is then
unconditionally set to the value of `GfwEnableFsCache`. This ensures that
even if a user attempts to circumvent the steps above, `ENABLEFSCACHE`
always contains a value based on the order of precedence listed above.

Additionally, if the user is performing a full UI install and the user
chooses to disable this feature, `ENABLEFSCACHE` is set to null (`{}` in
MSI parlance). MSI installer check boxes work like HTML check boxes, if
the `ENABLEFSCACHE` property has a value--any value--the check box will be
checked, enabling the feature.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>