msi: Fix ENABLEGCM property
While initially implementing the MSI properties, this developer made an
assumption that the default values were initially used for an installation
in order to test that upon upgrade the property would be set to the
original installation value. In addition, this developer tested setting
the property on the command line--but again, under the assumption that the
default value was initially used.

Unfortunately, this developer never took the implementation to the next
step, where the initial installation used a non-default value--and how
the property implementation responded in light of that.

This commit fixes the ENABLEGCM property so that upgrade and
repair/reinstall scenarios work correctly, especially when the installed
value for the property is initially a non-default value _and_ the property
value is overridden on the command line for one of these installation
scenarios.

It turns out that the implementation of the Remember Property pattern as
shown on @robmen's blog is only a very basic implementation; and as such,
it is a good example demonstrating a specific problem that may come across
when developing an MSI installer. However, since it's only a simple
example, it's not particularly robust; and in the end, this developer
determined that a different implementation was needed to best support the
particular goals of this MSI.

The problem with the implementation as shown at @robmen's blog is that for
more complex scenarios (such as for the Git for Windows installer), it
becomes impossible to tell if an installer property is set from the
command line, was detected as the value used during a previous install
during an upgrade, or the current value of the property is the default
value. This makes supporting upgrade, repair, and reinstall scenarios
nearly impossible.

This new approach to the Remember Property pattern supports upgrade,
repair, and reinstall scenarios while supporting the correct hierarchy of
property value precedence (in order of highest precedence to lowest
precedence):

- The value selected through the installer UI
- The value provided on the command line invocation of the installer
- The value set from a previous installation (i.e. upgrade install)
- The property's default value (i.e. fresh install)

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>