msi: a fix when setting CLIENVIRONMENT="Bash"
Back when PR #113 was merged, a mis-spelled variable name slipped through
review. When the user chooses the CLI environment to be used with Git for
Windows, if the user choses to only use Bash with Git for Windows, the
variable `CLIENVIRONMENTPATHS` should be cleared so that the user's `PATH`
system environment variable will not be modified (i.e. Git will not be
able to be used from the Windows command prompt). This fix corrects the
spelling of the variable to ensure that the user's `PATH` system
environment variable will not be modified.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>