msi: fix ENABLEFSCACHE and ENABLEGCM checkboxes
As this developer has mentioned many times, MSI checkboxes work like HTML
checkboxes. If the property backing an MSI checbox has any value, the
checkbox is checked--and this would translate into an enabled property.

This developer assumed that once the user clicks `Next` in a dialog, the
user is finished with that dialog, and so if a checkbox is unchecked, the
property is normally deleted; but instead, the property was reset with the
"disabled" value (`false` or `0`). However, if the user decides to click
`Back` to return to this dialog, now the property has a value and so the
checkbox will appeared checked. Furthermore, because the checkbox is
checked, upon clicking `Next`, the feature will now be enabled.

Instead of setting the property's `false` value in the dialog, let the
property remain deleted. Then, before the `InstallValidate` action, make
sure these properties, if they don't exist, contain a `false` value to
disable the feature.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>