please.sh: offer a function to rebase our master to upstream
Every once in a while, upstream releases a new version, and Git for
Windows' `master` branch is rebased with the `shears` script (to preserve
the branch structure).

This strategy implies that every once in a while, a terribly difficult
merge conflict and/or test failure will delay the Git for Windows release.

A better strategy is to rebase more often, of course. But that would
clutter the history of Git for Windows `master` branch, as the merging
rebases duplicate history.

But then, we do not really have to merge back the rebased branches at
all... We can simply perform merging rebases of Git for Windows' `master`
branch onto upstream's integration branches, just to determine whether
there are merge conflicts and/or test failures. And we can use `rerere` to
remember merge conflict resolutions.

So here is a commit that facilitates that workflow:

	please.sh rebase <branch>

This can be used e.g. with the `pu` branch, to catch problems early, and
quite possibly throwing them back to the patch authors whose patches broke
things, rather than this poor old maintainer having to resolve issues in a
hurry on release days.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>