please.sh rebase: allow continuing a rebase
The real problem with rebasing Git for Windows' `master` branch to any
upstream branch is not the rebasing per se, but to resolve any conflicts
while they arise.

The `rebase` subcommand already runs the shears script with
rerere.enabled=true to record conflicts. However, when continuing with a
plain `git rebase --continue`, that config setting is no longer set
(unless the config was modified manually to that end).

So let's teach the `rebase` subcommand to continue the rebase after the
user resolved conflicts.

While at it, let's also record the conflict resolutions in a separate
branch: one of the intended usages of the `rebase` subcommand is to allow
us to use Continuous Integration to catch problems early, i.e. to run said
subcommand non-interactively, in a setup where we cannot teach rerere
interactively.

To allow transporting the merge conflict resolution, we now also craft
commits on top of a newly-introduced branch called `rerere-train`: the
intention is to feed those merge commits to the contrib/rerere-train.sh
script to prime `git rerere`. The new commits on `rerere-train` reflect a
merge between stages 2 and 3, based on stage 1, of the index when
continuing the rebase, with the merge result reflecting the worktree.

To allow this, the user *must not* call `git add -u` before asking the
please.sh script to continue the rebase, of course, but leave that task to
the script.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>