please.sh rebase: train `rerere` automatically
When calling the `rebase` subcommand, it is convenient to make use of
the previously recorded resolution of merge conflicts.

In the previous commit, the `rebase` subcommand learned to craft
commits on the newly-introduced `rerere-train` branch, to allow
transporting the knowledge about merge conflict resolutions to other
setups.

With this change, we push the `rerere-train` branch when it gets
updated, and also train rerere when we get updates via the remote
`rerere-train` branch.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>