please.sh: allow build agents to use tokens instead of SSH keys
Git has the option to pass additional HTTP headers when fetching or
pushing, and build agents can benefit from that: rather than using
full-fledged SSH keys (which may even be hard to transport safely and
securely to build agents), we can set

	GIT_CONFIG_PARAMETERS='http.<url>.extraheader=...'

via the secure environment variables that every major Continuous
Integration system supports. That makes the authentication more secure
because the environment variables are not stored on disk, and the tokens
can be generated with a more focused scope.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>