please.sh rebase/test_remote_branch: improve test run
This commit consolidates running the tests into a single function, and
improves it by re-running the failed tests (if any) in super-verbose
mode.

Incidentally, this re-run may let the tests succeed if they failed
previously due to resource shortage. Unfortunately also if the tests
themselves are racy and fail only occasionally, but that's kinda okay
because this developer has already enough work on his plate with
reproducible bugs.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>