please.sh rebase/test_remote_branch: make it work with build agents
Build agents typically run under special accounts, such as NETWORK
SERVICE (i.e. running as a Windows Service).

These accounts are not necessarily handled nicely by the Cygwin runtime
(and hence the MSYS2 runtime): NETWORK SERVICE, for example, is reported
to have no login shell by default. The Cygwin/MSYS2 runtime assigns the
(non-existing) /sbin/nologin as a consequence.

This is usually not a problem, because nobody is insane enough to try to
log in via those accounts. However, Git's test suite (t3404, to be
precise), resets SHELL to the empty string and expects the login shell
(that is substituted as a consequence) to be a valid shell. It is not:
it does not exist in above-mentioned case. That means that rebase -i
fails to run the rather trivial ">touch-one" command, and everything
falls apart.

We do not want everything to fall apart.

So let's just force the login shell of the current, special account to
/usr/bin/bash if we detect that there is a possibility for things to
fall apart.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>