please.sh rebase/test_remote_branch: skip test that cannot pass
Right now, long paths are not handled by Git for Windows. And t3200
tests implicitly for a very long path (which is however not the purpose,
the test just assumes that long branch names are not a problem for the
filesystem backend).

This is not a problem when running in C:/git-sdk-64/usr/src/git, but it
is in C:/git-sdk-64-ci/usr/src/MINGW-packages/mingw-w64-git/src/git (the
location where build agents run the tests).

Let's just skip this test case, specifically, for now.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>