please.sh: avoid using the name 'path' for variables
On Windows, environment variable names are not really case-sensitive:
while they are stored in a case-preserving way, retrieval is
case-insensitive.

The POSIX emulation layer provided by MSYS2 (and used by the Bash in Git
for Windows) provides case-sensitive lookup of environment variables,
though, and converts them to all upper-case when spawning a non-MSYS2
program, such as `git.exe`.

Therefore, using a variable with the name `path` (all lower-case) must
wreak havoc, depending in which order `PATH` and `path` are exported.

To avoid this problem, let's simply use a different variable name.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>