please.sh rebase/test_remote_branch: optionally override worktree
By default, the `rebase` and `test_remote_branch` commands re-use the
Git checkout of MINGW-packages' mingw-w64-git/ subdirectory.

However, in particular in the case of using those commands in a build
agent, it may make much more sense to use a separate worktree.

Let's allow that with the newly-introduced --worktree=<dir> option.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>