please.sh rebase/test_remote_branch: do not skip t3200.32 anymore
The problem with this work-around was that the different integrations
branches of upstream Git have the offending test case in different
positions (as of time of writing, `maint` has it as t3200.31 while
the other three have it as t3200.32).

We circumvented this problem by using the --worktree=<dir> option to use
a location that does not result in that insanely long prefix, so that
the insanely long branch name does not result in an absolute path that
crashes right into Windows' path limit.

So let's just revert the previous work-around, as it is no longer
needed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>