please.sh build git: use the signtool alias when available
In https://github.com/git-for-windows/MINGW-packages/commit/4ddc5d6 we
changed the process to build the mingw-w64-git packages to allow for
code-signing by setting the SIGNTOOL environment variable to an
appropriate command-line that just needs the file names as additional
arguments.

Since we already have code in please.sh that expects an optional signtool
configuration as alias, let's just reuse that, if available.

This helps in environments where Windows File Protection is enabled, which
suffers greatly when being used with unsigned .exe files (everything gets
super slow).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>