installer: respect saved PATH option again
For a while, we ignored previous settings of the PATH option, to allow the
new default to add /cmd/ to the PATH to trickle down to end users.

This has been in effect for almost five months and eight Git for Windows
releases. So it is finally time to respect previous settings again.

This partially reverts commit 1e2e00eeb(installer: add git.exe to the PATH
by default, ignoring previous settings).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>