installer: respect saved FSCache setting again
For a while, we ignored previous settings of the FSCache option, to
allow the new default to enable the fscache to trickle down to end
users.

This has been in effect for almost five months and eight Git for Windows
releases. So it is finally time to respect previous settings again.

This partially reverts commit a1ae1463a0(installer: enable FSCache by
default, ignoring previous settings).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>