git-extra: fix up Git's builtins
To "save space" (i.e. to stop users from complaining that their Explorer
shows 250MB usage in mingw64\libexec\git-core because hard linked files
are counted incorrectly), we drop in the git-wrapper.exe in place of the
builtins.

Leading up to Git for Windows v2.10.0, this logic (to overwrite the
hard-linked builtins) has been moved out from mingw-w64-git in favor of
being "fixed up" by git-extra's post-install script.

This is the second half of that change.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>