installer: switch to using edit-git-bash.exe
... instead of edit-git-bash.dll.

Previously, we required the .dll and let InnoSetup open it directly.
This required the .dll to be 32-bit because InnoSetup is 32-bit.

Cross-compiling was a hassle, though, and it would probably eventually
be a good idea anyway to ship mingw-w64-git with the utility that lets
you edit the command-line to launch for git-bash.exe.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>