please.sh: add a command to build a pre-release
A pre-release is not exactly a full, official installer, but can be built
from an arbitrary revision of Git for Windows' repository.

To this end, `please.sh` has to be called from a Git repository other than
/usr/src/MINGW-packages/mingw-w64-git/src/git/, and a specific revision
has to be provided (which can be `HEAD`).

The script will then create a tag, push that tag, and build custom
32-bit and 64-bit Pacman packages. It will then temporarily install those
packages into the 32-bit and 64-bit SDKs and build the installers.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>