please.sh prerelease: allow building MinGit
The MinGit package is special: it is a "Git for Windows applications". As
such, it makes sense to bundle it from an experimental Git revision just
for testing.

With this patch, we can do exactly that. In a Git checkout, call e.g.

	/usr/src/build-extra/please.sh prerelease --mingit HEAD

to build both 32-bit and 64-bit packages.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>