please.sh prerelease: temporarily edit the release notes
It would be confusing to read release notes that reference a different
version number and date.

Note: we still leave the rest of the release notes intact, even if it may
not apply to the particular Git revision we are building. It is the
responsibility of the caller to make sure that those release notes make
sense.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>