please.sh prerelease: improve error when packages were not installed
When temporarily installing package versions, we expect them to be
installed at a different version (and we will reinstall that version
after we're done).

Of course, we could simply uninstall the packages afterwards if they
were not installed in the first place. But that would complicate the
code, and we can easily accommodate the script's expectation.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>