please.sh prerelease: smartly reinstall previous Git versions
When building prereleases, we temporarily install the desired Git version,
and immediately revert back to the previous version.

And here lies the rub; if the previous version was installed via `pacman
-U <file>` instead of `pacman -S`, then the Pacman cache won't have the
necessary files.

But all is not lost! We may have just been installing Git versions built
on the very same machine as this prerelease; in this case, the files are
available in the build directory and we can simply copy them.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>