please.sh sync: stop hard-coding core packages
Recent MSYS2 pacman versions have an ugly hack to update core packages:
when core packages such as mintty, bash, etc need to be updated, *only*
core packages are updated and pacman stops. In interactive sessions, it
even pauses indefinitely, asking the user to force-close the terminal
window.

At least in non-interactive sessions (such as Continuous Integration
tasks), pacman exits after such an upgrade and returns execution to the
caller. This gives us a chance to inspect the log whether there might be
remaining updates, and to re-run pacman in such a case.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>