installer: skip querying pointless process information
When determining which processes still use .dll files we intend to
override, we go out of our way to determine the paths of their
corresponding executables. And then we never use those paths.

Let's just skip that step.

This also helps with enumerating ssh-agent.exe processes that cannot be
opened with PROCESS_VM_READ nor with PROCESS_QUERY_INFORMATION (for
security reasons).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>